+++ src/program.h
@@ -439,7 +439,7 @@ INT32 define_function(struct pike_string *name,
 int really_low_find_shared_string_identifier(struct pike_string *name,
 					     struct program *prog,
 					     int flags);
-int low_find_lfun(struct program *p, int lfun);
+int low_find_lfun(struct program *p, ptrdiff_t lfun);
 int low_find_shared_string_identifier(struct pike_string *name,
 				      struct program *prog);
 struct ff_hash;
